% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_intervals.R
\name{summary.error_interval}
\alias{summary.error_interval}
\title{Error Intervals Summaries}
\usage{
summary.error_interval(object, ...)
}
\arguments{
\item{object}{object of class \code{error_interval} to be printed.}

\item{...}{optional arguments.}
}
\value{
Object of class \code{c("summary.error_interval","list")} corresponding
to the summary of x.
}
\description{
\code{summary} produces summaries for
objects of class \code{error_interval}.
}
\examples{
res<-error_interval(rnorm(10))
summary(res)
}
\author{
Jesus Prada, \email{jesus.prada@estudiante.uam.es}
}
\references{
Link to the scientific paper

Prada, Jesus, and Jose Ramon Dorronsoro. "SVRs and Uncertainty Estimates in Wind 
Energy Prediction." Advances in Computational Intelligence. Springer International 
Publishing, 2015. 564-577,

with theoretical background for this package is provided below.

\url{http://link.springer.com/chapter/10.1007/978-3-319-19222-2_47}
}
\seealso{
\link{error_interval}
}

